requirements.txt
typer
rich
fastapi
uvicorn
requests
openai
groq
python-dotenv
